# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/help/faq": {
      "get": {
        "operationId": "HelpController_getFAQs",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of FAQs to return (default: 8)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query to filter FAQs by title or content",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "FAQs retrieved successfully"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Get FAQs from help_support table",
        "tags": [
          "Help & Support"
        ]
      }
    },
    "/api/help/ai": {
      "post": {
        "operationId": "HelpController_handleAIChat",
        "parameters": [],
        "responses": {
          "200": {
            "description": "AI response generated successfully"
          },
          "400": {
            "description": "Bad request (invalid message, rate limit, etc.)"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Send a message to the AI assistant",
        "tags": [
          "Help & Support"
        ]
      }
    },
    "/api/help/content": {
      "get": {
        "operationId": "HelpController_getAllHelpSupport",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category: FAQ, System Message, Policy",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of help support items retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Get all help support content (admin only)",
        "tags": [
          "Help & Support"
        ]
      },
      "post": {
        "operationId": "HelpController_createHelpSupport",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Help support item created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a new help support item (admin only)",
        "tags": [
          "Help & Support"
        ]
      }
    },
    "/api/help/content/{id}": {
      "put": {
        "operationId": "HelpController_updateHelpSupport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Help support item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Help support item updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Help support item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update an existing help support item (admin only)",
        "tags": [
          "Help & Support"
        ]
      },
      "delete": {
        "operationId": "HelpController_deleteHelpSupport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Help support item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Help support item deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Help support item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a help support item (admin only)",
        "tags": [
          "Help & Support"
        ]
      }
    },
...
    "/api/fields": {
      "get": {
        "operationId": "FieldsController_getFields",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of fields retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all fields of interest",
        "tags": [
          "Fields"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your Supabase access token"
      }
    },
...
  }
}